refactor: modal 컴포넌트 및 hook 분리
