+++ src/interfaces/modal.ts
@@ -2,16 +2,19 @@ import { Dispatch } from "react";
 
 export interface IModal {
   children: JSX.Element;
-  setModal: Dispatch<boolean>;
 }
 
 export interface IResultModal {
-  setModal: Dispatch<boolean>;
+  closeModal: () => void;
   setLoading: Dispatch<boolean>;
   resultId: string;
 }
 
 export interface ISymptomModal {
-  setModal: Dispatch<boolean>;
+  closeModal: () => void;
   select: number;
+}
+
+export interface ILoginModal {
+  closeModal: () => void;
 }
