+++ src/pages/main/bottomButtons/index.tsx
@@ -1,10 +1,9 @@
-import { Dispatch } from "react";
 import theme from "../../../lib/theme";
 import RoundButton from "../../../components/roundButton";
 import { useNavigate } from "react-router-dom";
 import { Buttons, ButtonBox } from "./index.style";
 
-const BottomButtons = ({ setModal }: { setModal: Dispatch<boolean> }) => {
+const BottomButtons = ({ openModal }: { openModal: () => void }) => {
   const navigate = useNavigate();
 
   return (
@@ -14,11 +13,7 @@ const BottomButtons = ({ setModal }: { setModal: Dispatch<boolean> }) => {
           빠른 진단 시작하기
         </RoundButton>
       </ButtonBox>
-      <ButtonBox
-        onClick={() => {
-          setModal(true);
-        }}
-      >
+      <ButtonBox onClick={openModal}>
         <RoundButton outline="none" backgroundColor={theme.color.blue} color={theme.color.grey_100}>
           나의 진단기록장 보기
         </RoundButton>
